fix: duplicate table error with constraint field
perf: merge duplicate field when duplicate base and table
